Conflicts:
	CMakeLists.txt

+++ benchmarks/sample_assignment_from_py.cc
@@ -59,7 +59,7 @@ size_t speedtest (size_t size, size_t iters, float alpha, float d)
 
     size_t bogus = 0;
     double total_cats = 0;
-    for (int i = 0; i < iters; ++i) {
+    for (size_t i = 0; i < iters; ++i) {
         total_cats += max(model.sample_assignments(size, rng));
     }
 
@@ -86,11 +86,11 @@ int main (int argc, char ** argv)
     std::cout << "size" << '\t' << "cats" << '\t' << "samples/sec";
     std::cout << " (alpha = " << alpha << ", d = " << d << ")\n";
 
-    int min_exponent = 3;
-    int max_exponent = 6;
-    for (int i = min_exponent; i <= max_exponent; ++i) {
-        int size = int(round(pow(10, i)));
-        int iters = 10000000 / size;
+    size_t min_exponent = 3;
+    size_t max_exponent = 6;
+    for (size_t i = min_exponent; i <= max_exponent; ++i) {
+        size_t size = size_t(round(pow(10, i)));
+        size_t iters = 10000000 / size;
         speedtest(size, iters, alpha, d);
     }
