+++ benchmarks/sample_from_scores.cc
@@ -36,7 +36,7 @@ size_t speedtest (size_t size, size_t iters)
 {
     rng_t rng;
     std::vector<float> scores(size);
-    for (int i = 0; i < size; ++i){
+    for (size_t i = 0; i < size; ++i){
         scores[i] = 10 * sample_unif01(rng);
     }
 
@@ -45,14 +45,14 @@ size_t speedtest (size_t size, size_t iters)
     int64_t time = -current_time_us();
 
     size_t bogus = 0;
-    for (int i = 0; i < iters; ++i) {
+    for (size_t i = 0; i < iters; ++i) {
         bogus = sample_from_scores_overwrite(rng, scores_copy);
         scores_copy = scores;
     }
 
     time += 2 * current_time_us();
 
-    for (int i = 0; i < iters; ++i) {
+    for (size_t i = 0; i < iters; ++i) {
         scores_copy = scores;
     }
 
@@ -72,10 +72,10 @@ int main()
 {
     std::cout << "size" << '\t' << "choices/us" << '\n';
 
-    int max_exponent = 15;
-    for (int i = 1; i < max_exponent; ++i) {
-        int size = 1 << i;
-        int iters = 10 << (max_exponent - i);
+    size_t max_exponent = 15;
+    for (size_t i = 1; i < max_exponent; ++i) {
+        size_t size = 1 << i;
+        size_t iters = 10 << (max_exponent - i);
         speedtest(size, iters);
     }
